requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
scipy
gunicorn